+++ sdk/spring/azure-spring-boot-samples/azure-spring-boot-sample-cosmos/src/main/java/com/azure/cosmos/CosmosSampleApplication.java
@@ -2,12 +2,14 @@
 // Licensed under the MIT License.
 package com.azure.cosmos;
 
+import com.microsoft.azure.spring.autoconfigure.cosmos.CosmosCredentialConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
 import org.springframework.util.Assert;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -23,6 +25,9 @@ public class CosmosSampleApplication implements CommandLineRunner {
     @Autowired
     private UserRepository repository;
 
+    @Autowired
+    private ApplicationContext applicationContext;
+
     public static void main(String[] args) {
         SpringApplication.run(CosmosSampleApplication.class, args);
     }
@@ -30,7 +35,7 @@ public class CosmosSampleApplication implements CommandLineRunner {
     public void run(String... var1) throws Exception {
         final User testUser = new User("testId", "testFirstName", "testLastName", "test address line one");
 
-        // Save the User class to Azure Cosmos database.
+        // Save the User class to Azure Cosmos DB database.
         final Mono<User> saveUserMono = repository.save(testUser);
 
         final Flux<User> firstNameUserFlux = repository.findByFirstName("testFirstName");
@@ -53,8 +58,36 @@ public class CosmosSampleApplication implements CommandLineRunner {
         final User result = optionalUserResult.get();
         Assert.state(result.getFirstName().equals(testUser.getFirstName()), "query result firstName doesn't match!");
         Assert.state(result.getLastName().equals(testUser.getLastName()), "query result lastName doesn't match!");
-
         LOGGER.info("findOne in User collection get result: {}", result.toString());
+
+        //  Switch keys
+        CosmosCredentialConfiguration bean =
+            applicationContext.getBean(CosmosCredentialConfiguration.class);
+        bean.switchToCredential();
+        LOGGER.info("Switch to credential.");
+
+        final User testUserUpdated = new User("testIdUpdated", "testFirstNameUpdated",
+            "testLastNameUpdated", "test address Updated line one");
+        final User saveUserUpdated = repository.save(testUserUpdated).block();
+        Assert.state(saveUserUpdated != null, "Saved updated user must not be null");
+        Assert.state(saveUserUpdated.getFirstName().equals(testUserUpdated.getFirstName()),
+            "Saved updated user first name doesn't match");
+
+        final Optional<User> optionalUserUpdatedResult = repository.findById(testUserUpdated.getId()).blockOptional();
+        Assert.isTrue(optionalUserUpdatedResult.isPresent(), "Cannot find updated user.");
+        final User updatedResult = optionalUserUpdatedResult.get();
+        Assert.state(updatedResult.getFirstName().equals(testUserUpdated.getFirstName()),
+            "query updated result firstName doesn't match!");
+        Assert.state(updatedResult.getLastName().equals(testUserUpdated.getLastName()),
+            "query updated result lastName doesn't match!");
+
+        bean.switchToKey();
+        LOGGER.info("Switch back to key.");
+        final Optional<User> userOptional = repository.findById(testUserUpdated.getId()).blockOptional();
+        Assert.isTrue(userOptional.isPresent(), "Cannot find updated user.");
+        Assert.state(updatedResult.getFirstName().equals(testUserUpdated.getFirstName()),
+            "query updated result firstName doesn't match!");
+        LOGGER.info("Finished key switch.");
     }
 
     @PostConstruct